+++ gaios/UserSettings/Controllers/ShowMnemonicsViewController.swift
@@ -1,7 +1,7 @@
 import Foundation
 import UIKit
 import gdk
-
+import greenaddress
 
 class ShowMnemonicsViewController: UIViewController {
 
@@ -30,18 +30,28 @@ class ShowMnemonicsViewController: UIViewController {
             self.collectionView.reloadData()
             return
         }
+        let isHW = AccountsRepository.shared.current?.isHW ?? false
+        let derivedAccount = AccountsRepository.shared.current?.getDerivedLightningAccount()
         Task {
             do {
-                self.credentials = try await WalletManager.current?.prominentSession?.getCredentials(password: "")
-                self.bip39Passphrase = credentials?.bip39Passphrase
-                if showBip85, let credentials = self.credentials {
-                    self.lightningMnemonic = try await WalletManager.current?.getLightningMnemonic(credentials: credentials)
+                if isHW {
+                    if !showBip85 {
+                        throw GaError.GenericError("No export mnemonic from HW")
+                    } else if let derivedAccount = derivedAccount {
+                        self.lightningMnemonic = try AuthenticationTypeHandler.getAuthKeyCredentials(forNetwork: derivedAccount.keychain).mnemonic
+                    }
+                } else {
+                    self.credentials = try await WalletManager.current?.prominentSession?.getCredentials(password: "")
+                    self.bip39Passphrase = credentials?.bip39Passphrase
+                    if showBip85, let credentials = credentials {
+                        self.lightningMnemonic = try WalletManager.current?.getLightningMnemonic(credentials: credentials)
+                    }
                 }
                 await MainActor.run {
                     self.collectionView.reloadData()
                 }
             } catch {
-                print(error)
+                showError(error)
             }
         }
     }