+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Navigation of all external web urls in app web view
 - Update Breez SDK to 0.2.10
 
+### Fixed
+- Show derived lightning mnemonic from Jade
+
 ## [4.0.21] - 2023-12-08
 
 ### Added